Refactor: extract multitenant collection of 'live' reports
To help clarify subsequent changes
